Convert Header.test to TS and update snapshots. Tests pass and compiles fine
